Install pgbouncer on each app server
This creates session pooling for the database server. We have seen
hundreds of 500 errors due to the DB server running out of connections.

This fix limits the number of connections (in a pool) from the app to
the db server. Each server can have maximal 10 open sessions to the
db. This can be configured in the pgbouncer.ini file

The app servers now connect locally to the pgbouncer, the only change
is the IP address of the db server in owncloud/config/config.php (now
127.0.0.1)