Change setting for Postgres Autovacuum
Postgres default settings only start auto vacuuming when 20% of the
tuples in a table have become stale. By lowering this setting to 0.5% we
force vacuum much earlier. As some of the tables in the owncloud
database have millions of rows, we kept a large number of stale data in
the tables.